Adiciona tratamento de erros específico por tipo. Configura ambiente de desenvolvimento local

+++ src/main/java/com/sistemaguincho/gestaoguincho/dto/GuinchoRequestDTO.java
@@ -3,13 +3,15 @@ package com.sistemaguincho.gestaoguincho.dto;
 import com.sistemaguincho.gestaoguincho.enums.Disponibilidade;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class GuinchoRequestDTO {
 
     private String modelo;
     private String placa;
     private String tipo;
-    private Double capacidade;
+    private BigDecimal capacidade;
     private Disponibilidade disponibilidade;
 
 }
